requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.6
aiosignal==1.3.2
annotated-types==0.7.0
APScheduler==3.11.0
attrs==25.3.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.1
DateTime==5.5
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.2.0
packaging==24.1
propcache==0.3.0
psutil==6.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.4
schedule==1.2.2
setuptools==78.1.1
six==1.17.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.6.3
tzdata==2024.2
tzlocal==5.2
Werkzeug==3.1.5
WTForms==3.1.2
yarl==1.18.3
zipp==3.20.2
zope.interface==7.1.0